# dyninst: reduce memory usage for scraping and symbolicator