--- a/pkg/dyninst/gosym/cli/symbol.go
+++ b/pkg/dyninst/gosym/cli/symbol.go
@@ -11,32 +11,120 @@ package main
 import (
 	"errors"
 	"fmt"
+	"io"
 	"os"
 	"strconv"
+
+	"github.com/spf13/cobra"
 
 	"github.com/DataDog/datadog-agent/pkg/dyninst/gosym"
 	"github.com/DataDog/datadog-agent/pkg/dyninst/object"
 )
 
-func run(binary string, pc uint64) error {
+func main() {
+	if err := rootCmd.Execute(); err != nil {
+		os.Exit(1)
+	}
+}
+
+var rootCmd = &cobra.Command{
+	Short: "PC symbolication utilities",
+}
+
+func init() {
+	rootCmd.AddCommand(addr2lineCmd)
+	rootCmd.AddCommand(listCmd)
+}
+
+var addr2lineCmd = &cobra.Command{
+	Use:   "addr2line <binary> <pc>",
+	Short: "Resolve a PC to function@file:line",
+	Args:  cobra.ExactArgs(2),
+	RunE: func(cmd *cobra.Command, args []string) (retErr error) {
+		pc, err := strconv.ParseUint(args[1], 0, 64)
+		if err != nil {
+			return fmt.Errorf("invalid pc: %w", err)
+		}
+		binary := args[0]
+		cmd.SilenceUsage = true
+		symtab, closeFn, err := openGoSymbolTable(binary)
+		if err != nil {
+			return err
+		}
+		defer func() { retErr = errors.Join(retErr, closeFn()) }()
+		locations := symtab.LocatePC(pc)
+		if len(locations) == 0 {
+			return fmt.Errorf("no location found for pc 0x%x", pc)
+		}
+		for _, location := range locations {
+			fmt.Printf("%s@%s:%d\n", location.Function, location.File, location.Line)
+		}
+		return nil
+	},
+}
+
+var listCmd = &cobra.Command{
+	Use:   "list <binary>",
+	Short: "List all functions in a Go binary",
+	Args:  cobra.ExactArgs(1),
+	RunE: func(cmd *cobra.Command, args []string) (retErr error) {
+		binary := args[0]
+		cmd.SilenceUsage = true
+
+		symtab, closeFn, err := openGoSymbolTable(binary)
+		if err != nil {
+			return err
+		}
+		defer func() { retErr = errors.Join(retErr, closeFn()) }()
+		for fn := range symtab.Functions() {
+			fmt.Printf("%s %#x-%#x\n", fn.Name, fn.Entry, fn.End)
+		}
+		return nil
+	},
+}
+
+// openGoSymbolTable opens the binary and constructs the Go symbol table.
+// The returned close function must be called to release resources.
+func openGoSymbolTable(binary string) (
+	_ *gosym.GoSymbolTable, _ func() error, retErr error,
+) {
+	var closers []io.Closer
+	cleanupFn := func() error {
+		var err error
+		for _, closer := range closers {
+			err = errors.Join(err, closer.Close())
+		}
+		return err
+	}
+	defer func() {
+		if retErr != nil {
+			retErr = errors.Join(retErr, cleanupFn())
+		}
+	}()
+
 	mef, err := object.OpenMMappingElfFile(binary)
 	if err != nil {
-		return err
+		return nil, nil, err
 	}
-	defer func() { err = errors.Join(err, mef.Close()) }()
+	closers = append(closers, mef)
+
 	moduledata, err := object.ParseModuleData(mef)
 	if err != nil {
-		return err
+		return nil, nil, err
 	}
+
 	goDebugSections, err := moduledata.GoDebugSections(mef)
 	if err != nil {
-		return err
+		return nil, nil, err
 	}
-	defer func() { err = errors.Join(err, goDebugSections.Close()) }()
+	closers = append(closers, goDebugSections)
+
 	goVersion, err := object.ReadGoVersion(mef)
 	if err != nil {
-		return err
+		_ = errors.Join(goDebugSections.Close(), mef.Close())
+		return nil, nil, err
 	}
+
 	symtab, err := gosym.ParseGoSymbolTable(
 		goDebugSections.PcLnTab.Data(),
 		goDebugSections.GoFunc.Data(),
@@ -47,31 +135,8 @@ func run(binary string, pc uint64) error {
 		goVersion,
 	)
 	if err != nil {
-		return err
+		return nil, nil, err
 	}
-	locations := symtab.LocatePC(pc)
-	if len(locations) == 0 {
-		return fmt.Errorf("no location found for pc 0x%x", pc)
-	}
-	for _, location := range locations {
-		fmt.Printf("%s@%s:%d\n", location.Function, location.File, location.Line)
-	}
-	return nil
-}
 
-func main() {
-	if len(os.Args) != 3 {
-		fmt.Println("Usage: symbol <binary> <pc>")
-		os.Exit(1)
-	}
-	pc, err := strconv.ParseUint(os.Args[2], 16, 64)
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-	err = run(os.Args[1], pc)
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
+	return symtab, cleanupFn, nil
 }