… pkg/dyninst/gosym/symtab_benchmark_test.go
… pkg/dyninst/gosym/symtab_test.go
… pkg/dyninst/integration_test.go
… pkg/dyninst/object/gosym.go
… pkg/dyninst/object/goversion.go
… pkg/dyninst/rcscrape/ir_generator.go
… pkg/dyninst/rcscrape/ir_generator_test.go
… pkg/dyninst/symdb/symdb.go