This is fallout from #incident-42623. Along the way, we remove the need to parse the elf symbol table in order to load the gosym table. We also add some other optimizations to make it more efficient to search the gosym table by name. See individual commits.

#### dyninst/rcscrape: use gosym instead of elf symbols

```
name           old time/op    new time/op    delta
GenerateIR-14    4.07ms ±11%    1.95ms ± 7%  -52.00%  (p=0.000 n=10+10)

name           old alloc/op   new alloc/op   delta
GenerateIR-14    7.37MB ± 0%    0.01MB ± 0%  -99.82%  (p=0.000 n=8+8)

name           old allocs/op  new allocs/op  delta
GenerateIR-14     32.6k ± 0%      0.2k ± 0%  -99.40%  (p=0.000 n=10+10)
```


#### dyninst/rcscrape: add a benchmark for ir generation


#### dyninst/object: add gosym parsing utilities

This code was getting repeated too much. As part of this, the GoVersion
symbol needed to move to the gosym package to invert the dependencies.

Later on, if we feel sad about object depending on gosym (I do not), we
can add a third package that puts them together.


#### dyninst/gosym: remove dependency on GoVersion

Fixes https://datadoghq.atlassian.net/browse/DEBUG-4442.


#### dyninst/gosym: lazily resolve function names


#### dyninst/object: remove stale TODO


#### dyninst/gosym/cli: adopt cobra, add list command

I wanted to look at the functions, this made it easier to do that.


#### dyninst/gosym: provide streaming iteration, fewer ptrs

It's nice to be able to iterate the functions without keeping
them all in RAM at the same time.

Along the way, I removed some allocations:

```
name                                      old time/op    new time/op    delta
ParseGoSymbolTable/ListFunctions-14          213µs ± 2%     141µs ± 2%  -34.00%  (p=0.000 n=10+10)
ParseGoSymbolTable/LocatePC-14               873ns ± 4%     763ns ± 3%  -12.66%  (p=0.000 n=9+10)

name                                      old alloc/op   new alloc/op   delta
ParseGoSymbolTable/ListFunctions-14          261kB ± 0%     203kB ± 0%  -22.22%  (p=0.002 n=8+10)
ParseGoSymbolTable/LocatePC-14                522B ± 0%      362B ± 0%  -30.70%  (p=0.000 n=10+10)

name                                      old allocs/op  new allocs/op  delta
ParseGoSymbolTable/ListFunctions-14          5.91k ± 0%     1.97k ± 0%  -66.72%  (p=0.000 n=10+10)
ParseGoSymbolTable/LocatePC-14                11.0 ± 0%       6.0 ± 0%  -45.45%  (p=0.000 n=10+10)
```